+++ src/x2net/Buffer.cs
@@ -443,7 +443,7 @@ namespace x2net
             int limit = (marker >= 0 ? marker : back);
             if ((position + numBytes) > limit)
             {
-                Trace.Warn("front={0} pos={1} back={2} marker={3} numBytes={4}", front, position, back, marker, numBytes);
+                Trace.Info("front={0} pos={1} back={2} marker={3} numBytes={4}", front, position, back, marker, numBytes);
 
                 throw new System.IO.EndOfStreamException();
             }