+++ src/x2net/Links/NamedPipe/NamedPipeServer.cs
@@ -86,7 +86,7 @@ namespace x2net
             }
             catch (Exception ex)
             {
-                Trace.Warn("{0} {1} accept error: {2}",
+                Trace.Info("{0} {1} accept error: {2}",
                     Name, session.InternalHandle, ex);
                 return false;
             }