+++ src/x2net/Links/Socket/BeginEnd/UdpLink.cs
@@ -83,7 +83,7 @@ namespace x2net
                     }
                 }
 
-                Trace.Warn("{0} recvfrom error {1}", Name, e);
+                Trace.Info("{0} recvfrom error {1}", Name, e);
             }
 
             BeginReceiveFrom();
@@ -113,7 +113,7 @@ namespace x2net
                     }
                 }
 
-                Trace.Warn("{0} sendto error {1}", Name, e);
+                Trace.Info("{0} sendto error {1}", Name, e);
             }
 
             OnSendToInternal(bytesTransferred);