+++ src/x2net/Links/Socket/SocketAsyncEventArgs/AsyncTcpServer.cs
@@ -100,7 +100,7 @@ namespace x2net
                     }
                     else
                     {
-                        Trace.Warn("{0} accept error : {1}", Name, e.SocketError);
+                        Trace.Info("{0} accept error : {1}", Name, e.SocketError);
                     }
                 }
             }
@@ -110,7 +110,7 @@ namespace x2net
             }
             catch (Exception ex)
             {
-                Trace.Warn("{0} accept error : {1}", Name, ex);
+                Trace.Info("{0} accept error : {1}", Name, ex);
             }
 
             AcceptImpl(e);  // chain into the next accept